requirements.txt
azure-iot-device==2.14.0
python-dotenv==1.2.2